+++ index.ts
@@ -4,4 +4,9 @@ export { PopoverQueue } from './PopoverQueue'
 export { Popover3DRenderer } from './Popover3DRenderer'
 export { Popover3DAnimator } from './Popover3DAnimator'
 
-export { POPOVER_CONFIG, Popover3DPositioningMode } from './PopoverConfig'
+export {
+  getConfig,
+  POPOVER_CONFIG,
+  Popover3DPositioningMode,
+  type PopoverRuntimeOverrideKey,
+} from './PopoverConfig'